# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/brokermr810/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/fast-analysis/analyze:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Fast AI analysis for any symbol.
      description: 'Request body:

        market (required): Crypto, USStock, Forex, etc.

        symbol (required): e.g. BTC/USDT, AAPL

        language (optional, default en-US): Response language

        model (optional): LLM model id, e.g. openai/gpt-5.4

        timeframe (optional, default 1D): Analysis timeframe

        async_submit (optional): Submit as background task'
      tags:
      - FastAnalysis
      operationId: postApiFastAnalysisAnalyze
      x-visibility: public
  /api/fast-analysis/history:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get analysis history for a symbol.
      description: GET /api/fast-analysis/history?market=Crypto&symbol=BTC/USDT&days=7&limit=10
      tags:
      - FastAnalysis
      operationId: getApiFastAnalysisHistory
      x-visibility: public
  /api/fast-analysis/history/all:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get all analysis history with pagination.
      description: GET /api/fast-analysis/history/all?page=1&pagesize=20
      tags:
      - FastAnalysis
      operationId: getApiFastAnalysisHistoryAll
      x-visibility: public
  /api/fast-analysis/history/{memory_id}:
    delete:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Delete a history record.
      description: DELETE /api/fast-analysis/history/123
      tags:
      - FastAnalysis
      operationId: deleteApiFastAnalysisHistoryMemoryId
      x-visibility: public
    parameters:
    - in: path
      name: memory_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/fast-analysis/feedback:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Submit user feedback on an analysis.
      description: 'Request body:

        memory_id (required): Analysis history ID

        feedback (required): helpful, not_helpful, accurate, or inaccurate'
      tags:
      - FastAnalysis
      operationId: postApiFastAnalysisFeedback
      x-visibility: public
  /api/fast-analysis/performance:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get AI analysis performance statistics.
      description: GET /api/fast-analysis/performance?market=Crypto&symbol=BTC/USDT&days=30
      tags:
      - FastAnalysis
      operationId: getApiFastAnalysisPerformance
      x-visibility: public
  /api/fast-analysis/similar-patterns:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get similar historical patterns for current market conditions.
      description: GET /api/fast-analysis/similar-patterns?market=Crypto&symbol=BTC/USDT
      tags:
      - FastAnalysis
      operationId: getApiFastAnalysisSimilarPatterns
      x-visibility: public
...
openapi: 3.0.3